Image Support
Image support has been added, though it is still buggy.
The document should appear properly when initialized, but the images and the text seem to have different unit scales and origins, and need to be scaled and offset separately.
At the moment I haven't corrected these differences in the scaling or image logic, so images move out of place when you scale the document.
I think I can rewrite the draw function to draw the images points in a different order, or possibly just add some more hacky offset tweaking to get it to look the way I know it should.
Still possible to lose a draw handler if the area is made fullscreen, possibly under other circumstances. Will need to revisit that in the future.
Image scaling is also not being handled, so an HD image will dominate the document, could maybe add a per image scale factor to the `images` dict.
Themes and UI still coming.